refactor(constraints): move versionsCreatedSince to ArtifactRepository
(cherry picked from commit e498c5ebe71ac4a5c8b19e239a9a5e1b77a599fe)
